Add further details to all entity views
