+++ pages/entities/[id].vue
@@ -1,8 +1,6 @@
 <script lang="ts" setup>
-import { groupByToMap, keyByToMap } from "@acdh-oeaw/lib";
 import { z } from "zod";
 
-import { useIdPrefix } from "@/composables/use-id-prefix";
 import { hasCoordinates } from "@/utils/has-geojson-coordinates";
 
 // defineRouteRules({
@@ -26,8 +24,6 @@ const t = useTranslations();
 usePageMetadata({
 	title: t("EntityPage.meta.title"),
 });
-
-const { getUnprefixedId } = useIdPrefix();
 
 const route = useRoute();
 const id = computed(() => {
@@ -76,19 +72,6 @@ const tabs = computed(() => {
 	return tabs;
 });
 
-const relationsByType = computed(() => {
-	return groupByToMap(entity.value?.relations ?? [], (relation) => {
-		// FIXME: This used to use `relationType` (without the prefix)
-		// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-		return relation.relationSystemClass!;
-	});
-});
-
-const typesById = computed(() => {
-	return keyByToMap(entity.value?.types ?? [], (type) => {
-		return type.identifier;
-	});
-});
 </script>
 
 <template>
@@ -126,65 +109,6 @@ const typesById = computed(() => {
 				</TabsContent>
 			</Tabs> -->
 
-			<!-- <Card>
-				<CardHeader>
-					<CardTitle>{{ t("EntityPage.details") }}</CardTitle>
-				</CardHeader>
-				<CardContent>
-					<dl
-						class="grid gap-x-8 gap-y-4 sm:grid-cols-[repeat(auto-fill,minmax(20rem,1fr))] sm:justify-start"
-					>
-						<div v-for="[relationType, relations] of relationsByType" :key="relationType">
-							<dt class="text-xs font-medium uppercase tracking-wider text-muted-foreground">
-								{{ t(`SystemClassNames.${relationType}`) }}
-							</dt>
-							<dd>
-								<ul role="list">
-									<li v-for="(relation, index) of relations.slice(0, 10)" :key="index">
-										<NavLink
-											class="underline decoration-dotted hover:no-underline"
-											:href="{ path: `/entities/${getUnprefixedId(relation.relationTo)}` }"
-										>
-											{{ relation.label }}
-										</NavLink>
-										<span
-											v-if="
-												relation.relationSystemClass === 'type' &&
-												typesById.has(relation.relationTo)
-											"
-										>
-											({{ typesById.get(relation.relationTo)?.hierarchy }})
-										</span>
-									</li>
-								</ul>
-								<details v-if="relations.length > 10">
-									<summary class="cursor-pointer py-1 text-sm text-muted-foreground">
-										Show more
-									</summary>
-									<ul role="list">
-										<li v-for="(relation, index) of relations.slice(10)" :key="index">
-											<NavLink
-												class="underline decoration-dotted hover:no-underline"
-												:href="{ path: `/entities/${getUnprefixedId(relation.relationTo)}` }"
-											>
-												{{ relation.label }}
-											</NavLink>
-											<span
-												v-if="
-													relation.relationSystemClass === 'type' &&
-													typesById.has(relation.relationTo)
-												"
-											>
-												({{ typesById.get(relation.relationTo)?.hierarchy }})
-											</span>
-										</li>
-									</ul>
-								</details>
-							</dd>
-						</div>
-					</dl>
-				</CardContent>
-			</Card> -->
 		</template>
 
 		<template v-else-if="isLoading">
